modules/engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
firebase-admin==6.5.0
pydantic[email]==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
langextract>=0.1.0
google-genai>=0.2.0
nltk>=3.8.0
